+++ packages/german-words-dict/example.js
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
 const GermanWords = require('./dist/words.json');
 
 console.log(GermanWords['Kartoffel']);
