+++ packages/german-words-dict/test/test.js
@@ -8,13 +8,14 @@ const assert = require('assert');
 const germanWords = require('../dist/words.json');
 
 describe('german-words-dict', function () {
-  it('should work', function (done) {
+  it('has some content', function () {
     assert(germanWords != null);
     assert(Object.keys(germanWords).length > 100);
+  });
+  it('Kartoffeln are ok', function () {
     const kartoffel = germanWords['Kartoffel'];
     assert(kartoffel != null);
     assert.strictEqual(kartoffel['G'], 'F');
     assert.strictEqual(kartoffel['DAT']['PLU'], 'Kartoffeln');
-    done();
   });
 });
