+++ packages/rosaenlg/src/LanguageGerman.ts
@@ -12,8 +12,9 @@ import { getDet as getGermanDet } from 'german-determiners';
 import { agreeGermanAdjective, DetTypes as GermanDetTypes } from 'german-adjectives';
 import { AdjectivesInfo as GermanAdjectivesInfo } from 'german-adjectives-dict';
 import germanAdjectivesDict from 'german-adjectives-dict/dist/adjectives.json';
-import { getCaseGermanWord, getGenderGermanWord, WordsInfo as GermanWordsInfo } from 'german-words';
-import germanWordsDict from 'german-words-dict';
+import { getCaseGermanWord, getGenderGermanWord } from 'german-words';
+import germanWordsDict from 'german-words-dict/dist/words.json';
+import { WordsInfo as GermanWordsInfo } from 'german-words-dict';
 import { getOrdinal as getGermanOrdinal } from 'german-ordinals';
 import 'numeral/locales/de';
 import { de as dataFnsDe } from 'date-fns/locale';
